+++ lib/stm32/l1/Makefile
@@ -38,6 +38,7 @@ ARFLAGS		= rcs
 OBJS		= crc.o desig.o flash.o rcc.o usart.o dma.o lcd.o
 OBJS		+= crc_common_all.o dac_common_all.o
 OBJS		+= dma_common_l1f013.o
+OBJS		+= flash_common_l01.o
 OBJS		+= gpio_common_all.o gpio_common_f0234.o
 OBJS		+= i2c_common_v1.o iwdg_common_all.o
 OBJS		+= pwr_common_v1.o pwr_common_v2.o rtc_common_l1f024.o
